requirements.txt
tqdm
pandas
matplotlib
opencv-python
tensorflow-gpu